initial kernel launch epic
