+++ src/notebook/constants.js
@@ -5,6 +5,7 @@ export const DONE_SAVING = 'DONE_SAVING';
 
 export const ERROR_KERNEL_NOT_CONNECTED = 'ERROR_KERNEL_NOT_CONNECTED';
 
+export const LAUNCH_KERNEL = 'LAUNCH_KERNEL';
 export const NEW_KERNEL = 'NEW_KERNEL';
 export const KILL_KERNEL = 'KILL_KERNEL';
 export const INTERRUPT_KERNEL = 'INTERRUPT_KERNEL';
