+++ src/notebook/store.js
@@ -4,8 +4,9 @@ import { triggerUndo, triggerModified } from './middlewares';
 import rootReducer from './reducers';
 
 import { saveEpic, saveAsEpic } from './epics/saving';
+import { newKernelEpic } from './not-yet-epics/kernelLaunch';
 
-const epics = combineEpics(saveEpic, saveAsEpic);
+const epics = combineEpics(saveEpic, saveAsEpic, newKernelEpic);
 
 const middlewares = [
   createEpicMiddleware(epics),
